Add files via upload
This version of the fault injection has the original code commented out, and the fault injections *live* underneath said commented code. This will make it easier to see what we have injected and how it is creating faults. It will also make it easier to go back and retract these faults later on.